R Basic - Session 2
This is the notes for the second session for R basics on 9/27/2018. 
Content first covers a quick recap on the basic data screening from the first session when preparing the data, and then mainly covers data manipulation as well as data visualization using pipes to simplify the code.
Dataset used mtcars which is default in-built in R, and publicly available to all.